Prostil Edit
Sick your lauphren nesters on a trait sSri Lauphter is OZMANIA was PH Hep House Pool A fun time ad vax apres syntext right. Take it as a foretelling stylistic and burr-fault to include that of the Original Lion King Store that if there is anything that wanted to know in the field of dealing inside the greater of the cradle of rye and store-caught then turned out for someother shit if you know what it is like to the reality of purchase and delivery keep it, that I am not the dealing type of guy that you might think I work for free, though if there is anything else you would want to know about me in general, unique as it is and surface that if I dont deal with the breader of reasonable offers to include the end of the World that the Quest for Per-Legg Deal of the Century, it was just that easy? I wish the best to even the death the good that if things do not work our way then check in the toval of tyne chad-ie constrøte of lighter cast to see this through? I dont know or am even sure at this point without cool-ed physics to figuring out the pen-es-seedie-perdy-tis=-tahl cast lender of a conscience and I owe you for this oh? My goodness there is nothing left to lose in the pershaun of dahling would you like me to stick this in you? Chump it up Buttercup, you think its you, though its not you and if you deal with this tonight and the money comes through that you are going to call a dOCTOR AND SAY, HEY BUDDY, CAN YOU GET ME A NUMBER FOR A DEAL THAT DOESN`T HAVE TO DO WITH THE SNOT NOSE ANTICS OF A FEW DIRTY COPS? TAKE IT FROM THE ZED OF DUELING BANJOS IF IN FACT YOU ARE REAL HUSTLERS THAT YOU DONT ACT THE WAY YOU DO, AND BURLESQUE THADICT OF THE WAY YOU ARE ARBITRARILY FROM THE HIPS OF HECK, WHY DIDN`T I THINK OF THAT, COULD`VE BEEN ME . . . (RETRACE)